pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-work-publishment-for-wechat-article"
version = "0.1.0"
description = "Automates publishing Markdown articles with media to WeChat Official Account drafts."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT License" }

authors = [
  { name = "Your Name", email = "your.email@example.com" },
]

maintainers = [
  { name = "Your Name", email = "your.email@example.com" },
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Communications :: Chat",
  "Topic :: Text Processing :: Markup :: Markdown",
  "Topic :: Utilities",
]

dependencies = [
  "requests>=2.25.0",
  "python-dotenv>=0.15.0",
  "markdown>=3.3.0",
  "python-frontmatter>=1.0.0",
  "PyYAML>=5.3",
  "openai>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/your-repo-name"
"Repository" = "https://github.com/yourusername/your-repo-name"

[project.scripts]
wechat-publish = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
python_files = "test_*.py"
addopts = "-v"
testpaths = ["tests"]
pythonpath = [".", "src"]
